修复了部分BUG

+++ include/message.h
@@ -4,4 +4,4 @@ using namespace std;
 
 bool MessageCheck(string plain);
 string MessageReload(bool proxy, string http, string https);
-bool MessageLimit(string plain, int qq_num, int group_num, bool admin);
+bool MessageLimit(string plain, int64_t qq_num, int64_t group_num, bool admin);
